Moved `getAccessibilityRepresentation` inside `BaseCoBadgeCard`
